+++ clients/admin-ui/src/features/integrations/configure-monitor/MonitorDatabasePicker.tsx
@@ -96,6 +96,7 @@ const MonitorDatabasePicker = ({
     manualPagination: true,
     data,
     columns,
+    columnResizeMode: "onChange",
   });
 
   return (