+++ clients/admin-ui/src/pages/settings/domain-records.tsx
@@ -75,6 +75,7 @@ const DomainRecordsPage: NextPage = () => {
     getCoreRowModel: getCoreRowModel(),
     columns,
     data,
+    columnResizeMode: "onChange",
   });
 
   return (